Function send_to_kindle::send_to_kindle

pub async fn send_to_kindle(
    username: &str,
    password: &str,
    f: &str,
    ext: &str,
    file_timeout: usize,
    url: &str,
    daemon: bool,
    debugging_mode: bool
) -> WebDriverResult<()>

Parameters:

username: The amazon username as str

password: The amazon password as str

f: A path to a file or directory to be uploaded as str

ext: A file format extension to filter the files to be uploaded as str

file_timeout: Seconds assigned to each file before uploading times out

url: The amazon url to be used

daemon: Whether to run the geckodriver daemon or not. If not, it will have to be run externally

debugging_mode: For development purposes, enables GUI on the browser, does not send the files and prompts for user input to end the process

Description: Send the files to the kindle app and devices associated with the username and password amazon account

Notes: It is recommended to use the Args::default or Args::new functions to ensure some sane defaults and easier setup. Please note that the f str should be selected either from the file or directory field of the Args struct. f is filtered based on the ext parameter and the a vector of strings is constructed and passed to send_files_to_kindle